pyproject.toml
[project]
name = "client-intelligence-mcp"
version = "1.0.0"
description = "MCP server providing company intelligence — stock data, financial news, and Gemini-powered analysis"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Philippe Heymans" }]

dependencies = [
    "mcp[cli]==1.13.1",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "google-genai>=1.0.0",
    "pydantic>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
